requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
Flask==3.1.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
requests==2.32.5
urllib3==2.5.0
Werkzeug==3.1.3
